Simplify enums and remove references to bufferOptions
