+++ src/main/kotlin/espressoLogcat/Enums.kt
@@ -1,15 +1,23 @@
 package espressoLogcat
 
-enum class Priority {
-    VERBOSE,
-    DEBUG,
-    INFO,
-    WARNING,
-    ERROR,
-    FATAL,
-    SILENT
+/**
+ * Priority, also known as level, to be used when querying logcat
+ * `abbr` is the first character of the string
+ */
+enum class Priority(val abbr: String) {
+    VERBOSE(abbr = "V"),
+    DEBUG(abbr = "D"),
+    INFO(abbr = "I"),
+    WARNING(abbr = "W"),
+    ERROR(abbr = "E"),
+    FATAL(abbr = "F"),
+    SILENT(abbr = "S")
 }
 
+/**
+ * Various output formats. These are equivalent to passing in `-v {OutputFormat}` when querying logcat.
+ * However, EspressoLogcat only queries logcat with `threadtime` and then formats the response to match
+ */
 enum class OutputFormat {
     BRIEF,
     LONG,
@@ -20,22 +28,3 @@ enum class OutputFormat {
     THREADTIME,
     TIME
 }
-
-enum class BufferOptions {
-    RADIO,
-    EVENTS,
-    MAIN,
-    SYSTEM,
-    CRASH,
-    ALL,
-    DEFAULT
-}
-
-data class LineData(
-    val date: String,
-    val pid: String,
-    val tid: String,
-    val level: String,
-    val tag: String,
-    val coreData: String
-)
